fix: re-raise account retry delay
the previous commit lowered the retry delay because i thought it was being used to retry fulfillments. however, it turned out that that delay was only being used for querying the account. setting too short a delay causes the integration tests to fail, so this re-raises the delay to 1 second minimum but makes it clear that that retry value is only for querying the account